Clear clippy --all-features --all-targets -D warnings
Eight targeted lint fixes plus broad rustfmt-style cleanup across
the crate. Lib clippy was already clean; running with
`--all-targets -D warnings` exposed compile errors in the
integration test binaries that pre-dated this session
(`NetRouter::start` was changed to `Option<JoinHandle>` in
d3ae1fb7 but the integration tests still `.await`-ed it).

Lint fixes (lib):

- `int_plus_one` (`ffi/mod.rs`): `len() + 1 <= cap` → `len() < cap`
- `manual_is_multiple_of` (`ffi/mod.rs`): use the
  `is_multiple_of` method introduced in 1.95.
- `unnecessary_cast` x4 (`shard/mapper.rs`): drop redundant
  `as u64` from already-`u64` packed-atomic shifts/masks.
- `let_unit_value` (`adapter/net/compute/migration_source.rs`):
  drop the `let _claim =` binding around the entry-API match —
  the body returns `()` already.
- `clone_on_copy` (`adapter/net/session.rs` test):
  `header.clone()` → `header` since `NetHeader: Copy`.
- `neg_cmp_op_on_partial_ord`
  (`adapter/net/behavior/loadbalance.rs`): explicit
  `#[allow]` with comment — the `!(total_weight > 0.0)` form
  is intentional NaN-safety established by the audit's #115 fix
  (clippy's `partial_cmp` suggestion would re-introduce the
  NaN-falls-through-`<= 0.0` hazard).

Test compile fixes:

- `tests/three_node_integration.rs`: four sites of
  `Option<JoinHandle>.await` → `if let Some(h) = ... { let _ =
  h.await; }`.
- `tests/ffi_poll_buffer.rs`: c-string literal
  (`b"...\0".as_ptr() as *const c_char` → `c"...".as_ptr()`).
- `tests/migration_integration.rs`: update
  `assert!(orch.buffer_event(...))` to compare against the
  new `BufferOutcome::Buffered` variant (#143 changed the
  return type from bool).

Plus broader rustfmt-style normalization across the crate
applied by the auto-formatter (~340 lines of churn, mostly
collapsing multi-line attribute arguments onto a single line
where they fit, and similar cosmetic re-flow).

+++ net/crates/net/src/shard/mapper.rs
@@ -177,17 +177,15 @@ impl ShardMetricsCollector {
         // desync. Saturating ops cap at u32::MAX inside the
         // pack window (~4 G calls / 4 s of accumulated latency),
         // which is far beyond any sane metrics tick.
-        let _ = self.push_latency.fetch_update(
-            AtomicOrdering::Relaxed,
-            AtomicOrdering::Relaxed,
-            |v| {
-                let count = (v >> 32) as u32;
-                let sum = (v & 0xFFFF_FFFF) as u32;
-                let new_count = count.saturating_add(1) as u64;
-                let new_sum = sum.saturating_add(latency_ns.min(u32::MAX as u64) as u32) as u64;
-                Some((new_count << 32) | new_sum)
-            },
-        );
+        let _ =
+            self.push_latency
+                .fetch_update(AtomicOrdering::Relaxed, AtomicOrdering::Relaxed, |v| {
+                    let count = (v >> 32) as u32;
+                    let sum = (v & 0xFFFF_FFFF) as u32;
+                    let new_count = count.saturating_add(1) as u64;
+                    let new_sum = sum.saturating_add(latency_ns.min(u32::MAX as u64) as u32) as u64;
+                    Some((new_count << 32) | new_sum)
+                });
         // Always increment — the cost is one fetch_add and the
         // counter only matters when the shard is draining. Cheaper
         // than branching on `self.draining.load()` in the hot path.
@@ -282,11 +280,11 @@ impl ShardMetricsCollector {
         // chance of catching the sum without the matching count
         // (or vice versa) the way two independent swaps did.
         let push_packed = self.push_latency.swap(0, AtomicOrdering::Relaxed);
-        let push_count = (push_packed >> 32) as u64;
-        let push_latency_sum = (push_packed & 0xFFFF_FFFF) as u64;
+        let push_count = push_packed >> 32;
+        let push_latency_sum = push_packed & 0xFFFF_FFFF;
         let flush_packed = self.flush_latency.swap(0, AtomicOrdering::Relaxed);
-        let flush_count = (flush_packed >> 32) as u64;
-        let flush_latency_sum = (flush_packed & 0xFFFF_FFFF) as u64;
+        let flush_count = flush_packed >> 32;
+        let flush_latency_sum = flush_packed & 0xFFFF_FFFF;
 
         let fill_ratio = if self.capacity > 0 {
             current_len as f64 / self.capacity as f64
@@ -2472,10 +2470,7 @@ mod tests {
             .expect_err("scale_up immediately after drain_specific must hit cooldown");
         match err {
             ScalingError::InCooldown => {} // expected
-            other => panic!(
-                "expected InCooldown after drain_specific, got {:?}",
-                other
-            ),
+            other => panic!("expected InCooldown after drain_specific, got {:?}", other),
         }
     }
